+++ src/main/kotlin/dev/tomlarcher/gitarborist/settings/GitArboristSettings.kt
@@ -7,7 +7,6 @@ import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import dev.tomlarcher.gitarborist.carry.CarryOverScope
 import dev.tomlarcher.gitarborist.carry.CarryOverSource
-import dev.tomlarcher.gitarborist.git.WorktreeOpenMode
 
 /** Application-level persistent global defaults for Git Arborist. */
 @State(
@@ -26,7 +25,7 @@ class GitArboristSettings : PersistentStateComponent<GitArboristSettings.State> 
 
     data class State(
         var defaultWorktreeDirectory: String = ".worktrees",
-        var defaultOpenMode: WorktreeOpenMode = WorktreeOpenMode.NewWindow,
+        var openAfterCreate: Boolean = true,
         var carryOverScope: CarryOverScope = CarryOverScope.Curated,
         var carryOverSource: CarryOverSource = CarryOverSource.MainWorktree,
         var copyIdeaDirectory: Boolean = true,